+++ dist.browser/types.d.ts
@@ -1,5 +1,5 @@
 /// <reference types="bn.js" />
-import { BN } from 'ethereumjs-util';
+import { BN } from 'sbr-util';
 export interface genesisStatesType {
     names: {
         [key: string]: string;